Started reworking the event system so we can get it to work.
Specifically, events are now made up out of two SCM values, the first of which is intended for storing a symbol describing the event, the second of which holds the relevant data, which may be whatever the creator and receiver of the event agree upon. For communication between the engine and scripts, I'm considering a rather baroque conversion mechanism using an automated type dispatch to get the symbol and to invoke a function to get the data, both of which may be rather cumbersome, but given the vast array of SDL event types will probably be cleaner than a gigantic switch statement.