Event usage implemented, seems to be working pretty much fine.
We do still need implementations for more event types and a nice automated dispatcher over the type to grab the data and type information to pass to GUILE, but other than that, we're fine. There does seem to be a small bug with SDL_KEYUP events, which seem to not properly report their unicode key data, causing the dispatcher to throw SCM_EOL all the time. This should be fixed sometime in the future, preferably before the first release, to allow folks to properly use these events.

Other than that and getting that dispatcher done, the only major things left before this starts approaching an alpha release might very well be text rendering (for windows and the like, allowing text messages), a few guile APIs (for menus and the like) and the saving/loading of global data from GUILE (with an associated API for nice, cooperative usage as opposed to hogging the variable).