Preliminary sequencification of everything (well, almost) is complete, thanks to some code I had lying around.
Anyway, these might be slightly slower, but as this sequence implementation is a few orders of magnitude faster than the countless copies of the piece of crap we had before, it should be just fine. Of course, more functions should be moved over to sequences wherever that is practical and makes sense.