Finalized font opening/closing.
Basically, I fixed open_font up a bit to not waste memory when close_font left the usual holes in the sequence. Still TODO in the whole text rendering business is the actual text rendering and of course some.... interesting bits of math to allow rendering of text in windows (without spilling over all over the place if one is not careful). Thing is, I'm not sure if the engine should support that or if I should leave it to GUILE, but I'll throw it into the engine for now since it requires lots of basic, basic information that GUILE shouldn't have to care about (like the exact maximum heights of fonts).