Added prototype sed-based convertor macros.
Basically, the idea is to have sed take care of the stuff the C preprocessor won't do: LISP-macro-like constructs in C.
Currently, this requires GNU sed because of my use of \U and \E (Both of which could be avoided if the typeinfo enum macros were renamed to TYPE_<type-as-typedeffed> instead of the all-uppercase versions. That however would be extremely unesthetic and trading lots of ugliness for a bit of ugliness we can avoid is relatively lame.

This still needs to be integrated into the build system, probably using some kind of customized suffix rule.