Got rid of manual conversion func typing, except for pointer types.
This required a bit of a hack in Makefile.am, which will probably not work with compilers that run the linker even if you do supply -c (or for whom -c means 'mangle my code infinitely' or something like that), i.e. probably only works with the gcc. Other than that, it required some changes across the board in code files, most of which result in having far less boilerplate code around.