Changed camera math a bit.
Instead of calculating what needs to be substracted from the usual coordinate to get the new coordinate of the tilegrid origin, the offsets
now are what needs to be ADDED to the usual coordinate. Also removed some superfluous comments. It seems SDL does handle offsets just fine,
my math was a bit off.