sequence_map fixes: default version no longer modifies the sequence passed to it.
There is a new sequence_map_destructive (aka the renamed old function), which does. The point of this edition being that
it allows you to quickly replace the values in a sequence. Keep in mind that the procedure passed to sequence_map_destructive
is responsible for cleaning up the old object.