Changed cycling menus to include procedures.
Unfortunately, this affects users of cycling menus, since it requires a change to the data format.
Previously, the data format was a mere list of strings, but that means that the cycling menu's choice proc
returns some kind of name to the key binding procedure invoking (decide), which makes it rather impossible to
use menus in any meaningful way.

Now, the format is a list of pairs, where the car of each pair is a name tag, and the cdr is a procedure to be invoked
if said option is chosen. The procedure is called with the dialog concerned as its only argument.