Added: named grids.
Also made changes to add the recent work in the scheme libraries to the default conf file, and to use named grids in the testing game.

The main advantage of named grids is that they allow code to work with named regions,
making it gridid-agnostic. This is primarily useful for game maps, where you'd just
like to place something, without having to know your neighbouring region's exact
id.